Change implementation of optimalorder
Instead of going through all permutations of operations in their
grouped form (i.e. when desirable, operations are combined),
optimal order is decided purely based on the complexity
of sequentially eliminating edges.
The optimal edge-sequence is then translated to a list of operations
acting one or more edges each if it makes sense.